Organize code as much as possible
I tried to organize the code to make it a little more readable. That
said I think we could improve this further. At the same time, it might
be best to leave this for later, as the quickest fix is to use location
react hooks, but we can't use those until we get react router upgraded
to a newer version.