Use ref inside of useEffect
We have an effect to call when the component is initially render,
re-rendered, and unmounted. To do that we pass in an empty array as the
second parameter to React.useEffect(). In the modalController this
caused an error because we were manipulating the state within that
effect but not passing it as a dependency (we were using the empty
array).

To resolve this, I created a ref and a helper function to update the
value within that ref. The help function also updates the state.

Because we are now using a mutable object within the React.useEffect()
function we no longer have issues or errors due to missing dependencies.